Moved unit tests to live with pad spatial tests and deleted old test file.  Fixed unit tests.  Changed function to work in metres rather than kilometres for consistency with neighbourhood processing.
