+++ bin/improver-create-grid-with-halo
@@ -51,11 +51,11 @@ def main():
                         "file defining the target grid with additional halo.")
     parser.add_argument('--halo_size_km', metavar='HALO_SIZE', default=162.,
                         type=float, help="Size of halo (in km) with which to "
-                        "pad the input grid.")
+                        "pad the input grid.  Default is 162 km.")
     args = parser.parse_args()
 
     cube = load_cube(args.input_file)
-    halo_cube = create_cube_with_halo(cube, args.halo_size_km)
+    halo_cube = create_cube_with_halo(cube, 1000.*args.halo_size_km)
     save_netcdf(halo_cube, args.output_file)
